+++ client/src/index.js
@@ -5,8 +5,8 @@ import GlobalStyle from './GlobalStyle';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-    <React.StrictMode>
+    <>
         <GlobalStyle />
         <App />
-    </React.StrictMode>
+    </>
 );
